Moving encounter_id to reports. Allows easy access to cached Hierarchy
